WooAgent/agent/requirements.txt
# Core dependencies
openai-agents>=0.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
requests>=2.28.0
pydantic>=2.0.0
